Refactor MzmlParser to remove SwaMe and Prognosticator additions
This adds a minimal plugin architecture to MzmlReader (see class comment) and introduces a new project, SwaMe.Pipeline, as the plug-in for SwaMe and Prognosticator.  It also makes Scan and Run extensible by parameterising the MzmlReader and using a factory pattern for concrete scan and run generation.

Changing Ms2Scans from ConcurrentBag to List changes the order of enumeration of the MS2 scans.  It appears that the results of at least some metrics depended on the order (they shouldn't - multi-threaded parsing introduces inherent uncertainties into the enumeration order).  I strongly recommend that metric authors re-run their tests on real data and investigate any discrepancies with previous results.

+++ SwaMe.Pipeline/Scan.cs
@@ -1,21 +1,21 @@
 ﻿#nullable enable
 
+using MzmlParser;
 using ProtoBuf;
-using System.Collections.Generic;
+using System;
 using System.IO;
 
-namespace MzmlParser
+namespace SwaMe.Pipeline
 {
-    public class Scan
+    public class Scan : IScan
     {
-
-        public Scan(bool cacheSpectraToDisk, string tempDirectory) 
+        public Scan(bool cacheSpectraToDisk, string? tempDirectory)
         {
             CacheSpectraToDisk = cacheSpectraToDisk;
             TempDirectory = tempDirectory;
         }
 
-        public Scan(bool cacheSpectraToDisk, int isolationWindowLowerOffset, int isolationWindowUpperOffset, double scanStartTime, int msLevel, int density, string tempDirectory)
+        public Scan(bool cacheSpectraToDisk, int isolationWindowLowerOffset, int isolationWindowUpperOffset, double scanStartTime, int msLevel, int density, string? tempDirectory)
         {
             CacheSpectraToDisk = cacheSpectraToDisk;
             IsolationWindowLowerOffset = isolationWindowLowerOffset;
@@ -55,25 +55,13 @@ namespace MzmlParser
         public int RTsegment { get; set; }
         public int Density { get; set; }
         public double ProportionChargeStateOne { get; set; }
-        public string TempDirectory { get; set;  }
+        public string? TempDirectory { get; set; }
 
         private Spectrum? m_Spectrum;
 
-        public string ScanId
-        {
-            get
-            {
-                return string.Format("{0}_{1}_{2}", MsLevel, Cycle, IsolationWindowTargetMz);
-            }
-        }
+        public string ScanId => string.Format("{0}_{1}_{2}", MsLevel, Cycle, IsolationWindowTargetMz);
 
-        private string TempFileName
-        {
-            get
-            {
-                return Path.Combine(TempDirectory, string.Format("Yamato_{0}.tempscan", ScanId));
-            }
-        }
+        private string TempFileName => Path.Combine(TempDirectory, string.Format("Yamato_{0}.tempscan", ScanId));
         public Spectrum? Spectrum
         {
             get
@@ -102,42 +90,38 @@ namespace MzmlParser
             }
         }
 
-        public int SpectrumXmlBase64Line { get; set; }
+        #region IDisposable Support
+        private bool disposedValue = false; // To detect redundant calls
 
-        public int SpectrumXmlBase64LinePos { get; set; }
-
-        public int SpectrumXmlBase64Length { get; set; }
-
-        public int SpectrumXmlBase64BinaryOffset { get; set; }
-
-        public void DeleteTempFile()
+        protected virtual void Dispose(bool disposing)
         {
-            File.Delete(TempFileName);
-        }
-    }
+            if (!disposedValue)
+            {
+                if (disposing)
+                {
+                    // TODO: dispose managed state (managed objects).
+                }
 
-    [ProtoContract]
-    public class Spectrum
-    {
-        [ProtoMember(1)]
-        public virtual IList<SpectrumPoint>? SpectrumPoints { get; set; }
-    }
+                if (CacheSpectraToDisk)
+                    File.Delete(TempFileName);
+                m_Spectrum = null;
 
-    public class ScanAndTempProperties
-    {
-        public ScanAndTempProperties(bool cacheSpectra, string tempDirectory)
-        {
-            Scan = new Scan(cacheSpectra, tempDirectory);
+                disposedValue = true;
+            }
         }
 
-        public Scan Scan { get; set; }
-        public string? Base64IntensityArray { get; set; }
-        public string? Base64MzArray { get; set; }
-        public bool IntensityZlibCompressed { get; set; }
-        public bool MzZlibCompressed { get; set; }
-        public int IntensityBitLength { get; set; }
-        public int MzBitLength { get; set; }
+        ~Scan()
+        {
+          // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+          Dispose(false);
+        }
+
+        public void Dispose()
+        {
+            // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+        #endregion
     }
-
-
 }
